chore(text) Separate arrow and vector handling
