+++ examples/gpu-tables/arrow-text-2d/arrow-text-picking.ts
@@ -93,7 +93,7 @@ export function createArrowTextPickingModel(
     shaderLayout: TEXT_SHADER_LAYOUT,
     bufferLayout: textModel.bufferLayout,
     attributes: {
-      ...textModel.arrowGPUTable!.attributes,
+      ...textModel.table!.attributes,
       expandedGlyphVertexData: textModel.expandedGlyphVertexData
     },
     instanceCount: textModel.instanceCount,
@@ -123,21 +123,21 @@ function drawArrowTextPickingBatches(
   pickingModel: Model,
   textModel: AttributeTextModel
 ): void {
-  const arrowBatches = textModel.arrowGPUTable?.batches || [];
+  const gpuBatches = textModel.table?.batches || [];
   for (const [batchIndex, renderBatch] of textModel.renderBatches.entries()) {
-    const arrowBatch = arrowBatches[batchIndex];
-    if (!arrowBatch) {
-      throw new Error('Arrow text picking requires aligned Arrow and glyph render batches');
+    const gpuBatch = gpuBatches[batchIndex];
+    if (!gpuBatch) {
+      throw new Error('Arrow text picking requires aligned GPU and glyph render batches');
     }
     pickingModel.setAttributes({
-      ...arrowBatch.attributes,
+      ...gpuBatch.attributes,
       expandedGlyphVertexData: renderBatch.expandedGlyphVertexData
     });
     pickingModel.setInstanceCount(renderBatch.glyphCount);
     pickingModel.draw(pickingPass);
   }
   pickingModel.setAttributes({
-    ...(textModel.arrowGPUTable?.attributes || {}),
+    ...(textModel.table?.attributes || {}),
     expandedGlyphVertexData: textModel.expandedGlyphVertexData
   });
   pickingModel.setInstanceCount(textModel.glyphLayout.glyphCount);
