+++ modules/text/src/text-2d/arrow-attributes/convert-arrow-text-to-attribute.ts
