+++ modules/text/src/text-2d/arrow-attributes/convert-arrow-text-to-dictionary.ts
