+++ modules/text/src/text-2d/atlas/font-atlas-manager.ts
