+++ modules/text/src/text-2d/atlas/lru-cache.ts
