+++ modules/text/src/text-2d/atlas/text-utils.ts
