+++ modules/text/src/text-2d/model-utils/text-shaders.ts
@@ -1,0 +1,606 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+import type {ShaderLayout} from '@luma.gl/core';
+
+export const GLYPH_OFFSETS_COLUMN = 'glyphOffsets';
+export const GLYPH_FRAMES_COLUMN = 'glyphFrames';
+export const GLYPH_INDICES_COLUMN = 'glyphIndices';
+export const GLYPH_CLIP_RECTS_COLUMN = 'glyphClipRects';
+export const EXPANDED_GLYPH_VERTEX_DATA = 'expandedGlyphVertexData';
+export const COMPACT_GLYPH_VERTEX_DATA = 'compactGlyphVertexData';
+export const COMPACT_GLYPH_VERTEX_BYTE_STRIDE = Uint32Array.BYTES_PER_ELEMENT * 2;
+export const EXPANDED_GLYPH_VERTEX_BYTE_STRIDE = Uint32Array.BYTES_PER_ELEMENT * 4;
+export const CLIPPED_EXPANDED_GLYPH_VERTEX_BYTE_STRIDE =
+  EXPANDED_GLYPH_VERTEX_BYTE_STRIDE + Int16Array.BYTES_PER_ELEMENT * 4;
+export const INVALID_DICTIONARY_INDEX = 0xffffffff;
+
+export const DEFAULT_ARROW_TEXT_SHADER_LAYOUT: ShaderLayout = {
+  attributes: [
+    {name: 'positions', location: 0, type: 'vec2<f32>', stepMode: 'instance'},
+    {name: GLYPH_OFFSETS_COLUMN, location: 1, type: 'vec2<i32>', stepMode: 'instance'},
+    {name: GLYPH_FRAMES_COLUMN, location: 2, type: 'vec4<u32>', stepMode: 'instance'}
+  ],
+  bindings: []
+};
+
+export const DEFAULT_CLIPPED_ARROW_TEXT_SHADER_LAYOUT: ShaderLayout = {
+  attributes: [
+    ...DEFAULT_ARROW_TEXT_SHADER_LAYOUT.attributes,
+    {name: GLYPH_CLIP_RECTS_COLUMN, location: 3, type: 'vec4<i32>', stepMode: 'instance'}
+  ],
+  bindings: []
+};
+
+export const DEFAULT_STORAGE_INDEXED_TEXT_SHADER_LAYOUT: ShaderLayout = {
+  attributes: [
+    {name: GLYPH_OFFSETS_COLUMN, location: 0, type: 'vec2<i32>', stepMode: 'instance'},
+    {name: GLYPH_INDICES_COLUMN, location: 1, type: 'vec2<u32>', stepMode: 'instance'}
+  ],
+  bindings: []
+};
+
+export const DEFAULT_DICTIONARY_STORAGE_TEXT_SHADER_LAYOUT: ShaderLayout = {
+  attributes: [],
+  bindings: []
+};
+
+export const DEFAULT_ARROW_TEXT_VS = `#version 300 es
+precision highp float;
+
+in vec2 positions;
+in ivec2 glyphOffsets;
+in uvec4 glyphFrames;
+
+out vec2 vTextureCoordinate;
+
+uniform sampler2D fontAtlasTexture;
+
+vec2 getCorner(int vertexIndex) {
+  if (vertexIndex == 0) return vec2(0.0, 0.0);
+  if (vertexIndex == 1) return vec2(1.0, 0.0);
+  if (vertexIndex == 2) return vec2(1.0, 1.0);
+  if (vertexIndex == 3) return vec2(0.0, 0.0);
+  if (vertexIndex == 4) return vec2(1.0, 1.0);
+  return vec2(0.0, 1.0);
+}
+
+void main() {
+  vec2 corner = getCorner(gl_VertexID % 6);
+  vec2 atlasSize = vec2(textureSize(fontAtlasTexture, 0));
+  vec4 glyphFrame = vec4(glyphFrames);
+  vec2 glyphOffset = vec2(glyphOffsets);
+  vec2 glyphSize = glyphFrame.zw;
+  vec2 glyphPosition = positions + (glyphOffset + corner * glyphSize) * 0.001;
+  gl_Position = vec4(glyphPosition, 0.0, 1.0);
+  vTextureCoordinate = (glyphFrame.xy + corner * glyphSize) / atlasSize;
+}
+`;
+
+export const DEFAULT_CLIPPED_ARROW_TEXT_VS = `#version 300 es
+precision highp float;
+
+in vec2 positions;
+in ivec2 glyphOffsets;
+in uvec4 glyphFrames;
+in ivec4 glyphClipRects;
+
+out vec2 vTextureCoordinate;
+
+uniform sampler2D fontAtlasTexture;
+
+vec2 getCorner(int vertexIndex) {
+  if (vertexIndex == 0) return vec2(0.0, 0.0);
+  if (vertexIndex == 1) return vec2(1.0, 0.0);
+  if (vertexIndex == 2) return vec2(1.0, 1.0);
+  if (vertexIndex == 3) return vec2(0.0, 0.0);
+  if (vertexIndex == 4) return vec2(1.0, 1.0);
+  return vec2(0.0, 1.0);
+}
+
+bool isGlyphVertexClipped(vec2 glyphVertexOffset, ivec4 clipRect) {
+  if (clipRect.z >= 0) {
+    float clipMinX = float(clipRect.x);
+    float clipMaxX = clipMinX + float(clipRect.z);
+    if (glyphVertexOffset.x < clipMinX || glyphVertexOffset.x > clipMaxX) {
+      return true;
+    }
+  }
+  if (clipRect.w >= 0) {
+    float clipMinY = float(clipRect.y);
+    float clipMaxY = clipMinY + float(clipRect.w);
+    if (glyphVertexOffset.y < clipMinY || glyphVertexOffset.y > clipMaxY) {
+      return true;
+    }
+  }
+  return false;
+}
+
+void main() {
+  vec2 corner = getCorner(gl_VertexID % 6);
+  vec2 atlasSize = vec2(textureSize(fontAtlasTexture, 0));
+  vec4 glyphFrame = vec4(glyphFrames);
+  vec2 glyphOffset = vec2(glyphOffsets);
+  vec2 glyphSize = glyphFrame.zw;
+  vec2 glyphVertexOffset = glyphOffset + corner * glyphSize;
+  vec2 glyphPosition = positions + glyphVertexOffset * 0.001;
+  gl_Position = isGlyphVertexClipped(glyphVertexOffset, glyphClipRects)
+    ? vec4(0.0)
+    : vec4(glyphPosition, 0.0, 1.0);
+  vTextureCoordinate = (glyphFrame.xy + corner * glyphSize) / atlasSize;
+}
+`;
+
+export const DEFAULT_ARROW_TEXT_FS = `#version 300 es
+precision highp float;
+
+uniform sampler2D fontAtlasTexture;
+uniform float textUsesSdf;
+uniform float textSdfThreshold;
+uniform float textSdfSmoothing;
+in vec2 vTextureCoordinate;
+out vec4 fragColor;
+
+void main() {
+  float sampledAlpha = texture(fontAtlasTexture, vTextureCoordinate).a;
+  float sdfAlpha = textSdfSmoothing > 0.0
+    ? smoothstep(
+        textSdfThreshold - textSdfSmoothing,
+        textSdfThreshold + textSdfSmoothing,
+        sampledAlpha
+      )
+    : step(textSdfThreshold, sampledAlpha);
+  float alpha = textUsesSdf > 0.5 ? sdfAlpha : sampledAlpha;
+  fragColor = vec4(1.0, 1.0, 1.0, alpha);
+}
+`;
+
+export const DEFAULT_STORAGE_INDEXED_TEXT_SOURCE = /* wgsl */ `
+@group(0) @binding(auto) var fontAtlasTexture : texture_2d<f32>;
+@group(0) @binding(auto) var fontAtlasTextureSampler : sampler;
+@group(0) @binding(auto) var<storage, read> textRowPositions : array<vec2<f32>>;
+@group(0) @binding(auto) var<storage, read> textRowColors : array<u32>;
+@group(0) @binding(auto) var<storage, read> textRowAngles : array<f32>;
+@group(0) @binding(auto) var<storage, read> textRowSizes : array<f32>;
+@group(0) @binding(auto) var<storage, read> textRowPixelOffsets : array<vec2<f32>>;
+@group(0) @binding(auto) var<storage, read> textRowClipRects : array<vec2<u32>>;
+@group(0) @binding(auto) var<storage, read> textRowGlyphStarts : array<u32>;
+@group(0) @binding(auto) var<storage, read> textGlyphFrames : array<vec4<f32>>;
+
+struct TextStorageStyleConfig {
+  constantColor : vec4<f32>,
+  constantAngleDegrees : f32,
+  constantSize : f32,
+  constantPixelOffset : vec2<f32>,
+  useRowColors : u32,
+  useRowAngles : u32,
+  useRowSizes : u32,
+  useRowPixelOffsets : u32,
+  hasClipRects : u32,
+  batchRowIndexBase : u32,
+  rowStorageIndexBase : u32,
+  _padding0 : u32,
+  sdfThreshold : f32,
+  sdfSmoothing : f32,
+  _padding1 : vec2<f32>,
+};
+
+@group(0) @binding(auto) var<uniform> textStorageStyleConfig : TextStorageStyleConfig;
+
+struct TextStorageRenderConfig {
+  glyphIndexBase : u32,
+  rowStart : u32,
+  rowEnd : u32,
+  _padding : u32,
+};
+
+@group(0) @binding(auto) var<uniform> textStorageRenderConfig : TextStorageRenderConfig;
+
+struct VertexInputs {
+  @builtin(vertex_index) vertexIndex : u32,
+  @builtin(instance_index) instanceIndex : u32,
+  @location(0) glyphOffsets : vec2<i32>,
+  @location(1) glyphIndices : vec2<u32>,
+};
+
+struct FragmentInputs {
+  @builtin(position) position : vec4<f32>,
+  @location(0) atlasUV : vec2<f32>,
+  @location(1) color : vec4<f32>,
+};
+
+fn getGlyphCorner(vertexIndex: u32) -> vec2<f32> {
+  let corners = array<vec2<f32>, 6>(
+    vec2<f32>(0.0, 0.0),
+    vec2<f32>(1.0, 0.0),
+    vec2<f32>(1.0, 1.0),
+    vec2<f32>(0.0, 0.0),
+    vec2<f32>(1.0, 1.0),
+    vec2<f32>(0.0, 1.0)
+  );
+  return corners[vertexIndex % 6u];
+}
+
+fn unpackLowInt16(word: u32) -> i32 {
+  return i32(word << 16u) >> 16;
+}
+
+fn unpackHighInt16(word: u32) -> i32 {
+  return i32(word) >> 16;
+}
+
+fn unpackClipRect(words: vec2<u32>) -> vec4<i32> {
+  return vec4<i32>(
+    unpackLowInt16(words.x),
+    unpackHighInt16(words.x),
+    unpackLowInt16(words.y),
+    unpackHighInt16(words.y)
+  );
+}
+
+fn isGlyphVertexClipped(glyphVertexOffset: vec2<f32>, clipRect: vec4<i32>) -> bool {
+  if (clipRect.z >= 0) {
+    let clipMinX = f32(clipRect.x);
+    let clipMaxX = clipMinX + f32(clipRect.z);
+    if (glyphVertexOffset.x < clipMinX || glyphVertexOffset.x > clipMaxX) {
+      return true;
+    }
+  }
+  if (clipRect.w >= 0) {
+    let clipMinY = f32(clipRect.y);
+    let clipMaxY = clipMinY + f32(clipRect.w);
+    if (glyphVertexOffset.y < clipMinY || glyphVertexOffset.y > clipMaxY) {
+      return true;
+    }
+  }
+  return false;
+}
+
+fn unpackTextColor(colorWord: u32) -> vec4<f32> {
+  return vec4<f32>(
+    f32(colorWord & 0xffu),
+    f32((colorWord >> 8u) & 0xffu),
+    f32((colorWord >> 16u) & 0xffu),
+    f32((colorWord >> 24u) & 0xffu)
+  ) / 255.0;
+}
+
+fn rotateTextOffset(offset: vec2<f32>, angleDegrees: f32) -> vec2<f32> {
+  let angleRadians = angleDegrees * 0.017453292519943295;
+  let rotation = mat2x2<f32>(
+    cos(angleRadians),
+    sin(angleRadians),
+    -sin(angleRadians),
+    cos(angleRadians)
+  );
+  return rotation * offset;
+}
+
+fn findRowIndex(glyphIndex : u32) -> u32 {
+  var low = textStorageRenderConfig.rowStart;
+  var high = textStorageRenderConfig.rowEnd;
+  loop {
+    if (low >= high) {
+      break;
+    }
+    let middle = (low + high) / 2u;
+    let nextRowGlyphStart = textRowGlyphStarts[middle + 1u];
+    if (nextRowGlyphStart <= glyphIndex) {
+      low = middle + 1u;
+    } else {
+      high = middle;
+    }
+  }
+  return low;
+}
+
+@vertex
+fn vertexMain(inputs: VertexInputs) -> FragmentInputs {
+  let corner = getGlyphCorner(inputs.vertexIndex);
+  let glyphFrame = textGlyphFrames[inputs.glyphIndices.x];
+  let glyphOffset = vec2<f32>(inputs.glyphOffsets);
+  let glyphSize = glyphFrame.zw;
+  let glyphVertexOffset = glyphOffset + corner * glyphSize;
+  let glyphIndex = textStorageRenderConfig.glyphIndexBase + inputs.instanceIndex;
+  let rowIndex = findRowIndex(glyphIndex);
+  let rowStorageIndex = rowIndex + textStorageStyleConfig.rowStorageIndexBase;
+  var angleDegrees = textStorageStyleConfig.constantAngleDegrees;
+  if (textStorageStyleConfig.useRowAngles != 0u) {
+    angleDegrees = textRowAngles[rowStorageIndex];
+  }
+  var textSize = textStorageStyleConfig.constantSize;
+  if (textStorageStyleConfig.useRowSizes != 0u) {
+    textSize = textRowSizes[rowStorageIndex];
+  }
+  let textScale = textSize / 32.0;
+  let styledGlyphVertexOffset = rotateTextOffset(glyphVertexOffset * textScale, angleDegrees);
+  var pixelOffset = textStorageStyleConfig.constantPixelOffset;
+  if (textStorageStyleConfig.useRowPixelOffsets != 0u) {
+    pixelOffset = textRowPixelOffsets[rowStorageIndex];
+  }
+  let glyphPosition =
+    textRowPositions[rowStorageIndex] + (styledGlyphVertexOffset + pixelOffset) * 0.001;
+  let clipRect = unpackClipRect(textRowClipRects[rowIndex]);
+  let isClipped =
+    textStorageStyleConfig.hasClipRects != 0u &&
+    isGlyphVertexClipped(glyphVertexOffset, clipRect);
+  var outputs : FragmentInputs;
+  outputs.position = select(
+    vec4<f32>(glyphPosition, 0.0, 1.0),
+    vec4<f32>(0.0),
+    isClipped
+  );
+  outputs.atlasUV = (glyphFrame.xy + corner * glyphSize) / vec2<f32>(textureDimensions(fontAtlasTexture));
+  outputs.color = textStorageStyleConfig.constantColor;
+  if (textStorageStyleConfig.useRowColors != 0u) {
+    outputs.color = unpackTextColor(textRowColors[rowStorageIndex]);
+  }
+  return outputs;
+}
+
+@fragment
+fn fragmentMain(inputs: FragmentInputs) -> @location(0) vec4<f32> {
+  let sampledAlpha = textureSample(fontAtlasTexture, fontAtlasTextureSampler, inputs.atlasUV).a;
+  var alpha = sampledAlpha;
+  if (textStorageStyleConfig.sdfThreshold >= 0.0) {
+    if (textStorageStyleConfig.sdfSmoothing > 0.0) {
+      alpha = smoothstep(
+        textStorageStyleConfig.sdfThreshold - textStorageStyleConfig.sdfSmoothing,
+        textStorageStyleConfig.sdfThreshold + textStorageStyleConfig.sdfSmoothing,
+        sampledAlpha
+      );
+    } else {
+      alpha = select(0.0, 1.0, sampledAlpha >= textStorageStyleConfig.sdfThreshold);
+    }
+  }
+  return vec4<f32>(inputs.color.rgb, inputs.color.a * alpha);
+}
+`;
+
+export const DEFAULT_DICTIONARY_STORAGE_TEXT_SOURCE = /* wgsl */ `
+@group(0) @binding(auto) var fontAtlasTexture : texture_2d<f32>;
+@group(0) @binding(auto) var fontAtlasTextureSampler : sampler;
+// Dictionary text keeps row styling in row buffers, shared glyph layout in
+// dictionary buffers, and uses instance_index for visible glyph occurrences.
+// There is deliberately no per-visible-glyph vertex/occurrence buffer.
+@group(0) @binding(auto) var<storage, read> textRowPositions : array<vec2<f32>>;
+@group(0) @binding(auto) var<storage, read> textRowColors : array<u32>;
+@group(0) @binding(auto) var<storage, read> textRowAngles : array<f32>;
+@group(0) @binding(auto) var<storage, read> textRowSizes : array<f32>;
+@group(0) @binding(auto) var<storage, read> textRowPixelOffsets : array<vec2<f32>>;
+@group(0) @binding(auto) var<storage, read> textRowClipRects : array<vec2<u32>>;
+// textRowDictionaryRecords[row] = (dictionary value index, row glyph start).
+// The extra terminal record stores rowCount/invalid and total visible glyphs.
+@group(0) @binding(auto) var<storage, read> textRowDictionaryRecords : array<vec2<u32>>;
+// textDictionaryGlyphRanges[value] is a half-open range into
+// textDictionaryGlyphRecords for one unique dictionary string.
+@group(0) @binding(auto) var<storage, read> textDictionaryGlyphRanges : array<vec2<u32>>;
+// textDictionaryGlyphRecords[glyph] = (packed i16 x/y layout offset, glyph id).
+// Repeated rows reuse these records instead of copying glyph offsets and ids.
+@group(0) @binding(auto) var<storage, read> textDictionaryGlyphRecords : array<vec2<u32>>;
+@group(0) @binding(auto) var<storage, read> textGlyphFrames : array<vec4<f32>>;
+
+struct TextStorageStyleConfig {
+  constantColor : vec4<f32>,
+  constantAngleDegrees : f32,
+  constantSize : f32,
+  constantPixelOffset : vec2<f32>,
+  useRowColors : u32,
+  useRowAngles : u32,
+  useRowSizes : u32,
+  useRowPixelOffsets : u32,
+  hasClipRects : u32,
+  batchRowIndexBase : u32,
+  rowStorageIndexBase : u32,
+  _padding0 : u32,
+  sdfThreshold : f32,
+  sdfSmoothing : f32,
+  _padding1 : vec2<f32>,
+};
+
+@group(0) @binding(auto) var<uniform> textStorageStyleConfig : TextStorageStyleConfig;
+
+struct TextDictionaryRenderConfig {
+  glyphIndexBase : u32,
+  rowStart : u32,
+  rowEnd : u32,
+  _padding : u32,
+};
+
+@group(0) @binding(auto) var<uniform> textDictionaryRenderConfig : TextDictionaryRenderConfig;
+
+struct VertexInputs {
+  @builtin(vertex_index) vertexIndex : u32,
+  @builtin(instance_index) instanceIndex : u32,
+};
+
+struct FragmentInputs {
+  @builtin(position) position : vec4<f32>,
+  @location(0) atlasUV : vec2<f32>,
+  @location(1) color : vec4<f32>,
+};
+
+fn getGlyphCorner(vertexIndex: u32) -> vec2<f32> {
+  let corners = array<vec2<f32>, 6>(
+    vec2<f32>(0.0, 0.0),
+    vec2<f32>(1.0, 0.0),
+    vec2<f32>(1.0, 1.0),
+    vec2<f32>(0.0, 0.0),
+    vec2<f32>(1.0, 1.0),
+    vec2<f32>(0.0, 1.0)
+  );
+  return corners[vertexIndex % 6u];
+}
+
+fn unpackLowInt16(word: u32) -> i32 {
+  return i32(word << 16u) >> 16;
+}
+
+fn unpackHighInt16(word: u32) -> i32 {
+  return i32(word) >> 16;
+}
+
+fn unpackClipRect(words: vec2<u32>) -> vec4<i32> {
+  return vec4<i32>(
+    unpackLowInt16(words.x),
+    unpackHighInt16(words.x),
+    unpackLowInt16(words.y),
+    unpackHighInt16(words.y)
+  );
+}
+
+fn isGlyphVertexClipped(glyphVertexOffset: vec2<f32>, clipRect: vec4<i32>) -> bool {
+  if (clipRect.z >= 0) {
+    let clipMinX = f32(clipRect.x);
+    let clipMaxX = clipMinX + f32(clipRect.z);
+    if (glyphVertexOffset.x < clipMinX || glyphVertexOffset.x > clipMaxX) {
+      return true;
+    }
+  }
+  if (clipRect.w >= 0) {
+    let clipMinY = f32(clipRect.y);
+    let clipMaxY = clipMinY + f32(clipRect.w);
+    if (glyphVertexOffset.y < clipMinY || glyphVertexOffset.y > clipMaxY) {
+      return true;
+    }
+  }
+  return false;
+}
+
+fn unpackTextColor(colorWord: u32) -> vec4<f32> {
+  return vec4<f32>(
+    f32(colorWord & 0xffu),
+    f32((colorWord >> 8u) & 0xffu),
+    f32((colorWord >> 16u) & 0xffu),
+    f32((colorWord >> 24u) & 0xffu)
+  ) / 255.0;
+}
+
+fn rotateTextOffset(offset: vec2<f32>, angleDegrees: f32) -> vec2<f32> {
+  let angleRadians = angleDegrees * 0.017453292519943295;
+  let rotation = mat2x2<f32>(
+    cos(angleRadians),
+    sin(angleRadians),
+    -sin(angleRadians),
+    cos(angleRadians)
+  );
+  return rotation * offset;
+}
+
+fn emptyFragmentInputs() -> FragmentInputs {
+  var outputs : FragmentInputs;
+  outputs.position = vec4<f32>(0.0);
+  outputs.atlasUV = vec2<f32>(0.0);
+  outputs.color = vec4<f32>(0.0);
+  return outputs;
+}
+
+fn findRowIndex(glyphIndex: u32) -> u32 {
+  // Convert a glyph occurrence index back to its row by searching row starts.
+  // This replaces the older per-glyph row-index buffer.
+  var low = textDictionaryRenderConfig.rowStart;
+  var high = textDictionaryRenderConfig.rowEnd;
+  loop {
+    if (low >= high) {
+      break;
+    }
+    let middle = (low + high) / 2u;
+    let nextRowGlyphStart = textRowDictionaryRecords[middle + 1u].y;
+    if (nextRowGlyphStart <= glyphIndex) {
+      low = middle + 1u;
+    } else {
+      high = middle;
+    }
+  }
+  return low;
+}
+
+@vertex
+fn vertexMain(inputs: VertexInputs) -> FragmentInputs {
+  // Each instance is one visible glyph occurrence. The occurrence is not stored;
+  // glyphIndex plus the row/dictionary ranges resolves the shared glyph record.
+  let glyphIndex = textDictionaryRenderConfig.glyphIndexBase + inputs.instanceIndex;
+  let rowIndex = findRowIndex(glyphIndex);
+  if (rowIndex >= textDictionaryRenderConfig.rowEnd) {
+    return emptyFragmentInputs();
+  }
+  let rowDictionaryRecord = textRowDictionaryRecords[rowIndex];
+  let rowGlyphStart = rowDictionaryRecord.y;
+  let rowGlyphEnd = textRowDictionaryRecords[rowIndex + 1u].y;
+  if (glyphIndex < rowGlyphStart || glyphIndex >= rowGlyphEnd) {
+    return emptyFragmentInputs();
+  }
+  let dictionaryGlyphOffset = glyphIndex - rowGlyphStart;
+  let dictionaryIndex = rowDictionaryRecord.x;
+  if (dictionaryIndex == ${INVALID_DICTIONARY_INDEX}u) {
+    return emptyFragmentInputs();
+  }
+  let dictionaryGlyphRange = textDictionaryGlyphRanges[dictionaryIndex];
+  let dictionaryGlyphIndex = dictionaryGlyphRange.x + dictionaryGlyphOffset;
+  if (dictionaryGlyphIndex >= dictionaryGlyphRange.y) {
+    return emptyFragmentInputs();
+  }
+
+  let corner = getGlyphCorner(inputs.vertexIndex);
+  let glyphRecord = textDictionaryGlyphRecords[dictionaryGlyphIndex];
+  let glyphFrame = textGlyphFrames[glyphRecord.y];
+  let glyphOffset = vec2<f32>(
+    f32(unpackLowInt16(glyphRecord.x)),
+    f32(unpackHighInt16(glyphRecord.x))
+  );
+  let glyphSize = glyphFrame.zw;
+  let glyphVertexOffset = glyphOffset + corner * glyphSize;
+  let rowStorageIndex = rowIndex + textStorageStyleConfig.rowStorageIndexBase;
+  var angleDegrees = textStorageStyleConfig.constantAngleDegrees;
+  if (textStorageStyleConfig.useRowAngles != 0u) {
+    angleDegrees = textRowAngles[rowStorageIndex];
+  }
+  var textSize = textStorageStyleConfig.constantSize;
+  if (textStorageStyleConfig.useRowSizes != 0u) {
+    textSize = textRowSizes[rowStorageIndex];
+  }
+  let textScale = textSize / 32.0;
+  let styledGlyphVertexOffset = rotateTextOffset(glyphVertexOffset * textScale, angleDegrees);
+  var pixelOffset = textStorageStyleConfig.constantPixelOffset;
+  if (textStorageStyleConfig.useRowPixelOffsets != 0u) {
+    pixelOffset = textRowPixelOffsets[rowStorageIndex];
+  }
+  let glyphPosition =
+    textRowPositions[rowStorageIndex] + (styledGlyphVertexOffset + pixelOffset) * 0.001;
+  let clipRect = unpackClipRect(textRowClipRects[rowIndex]);
+  let isClipped =
+    textStorageStyleConfig.hasClipRects != 0u &&
+    isGlyphVertexClipped(glyphVertexOffset, clipRect);
+  var outputs : FragmentInputs;
+  outputs.position = select(
+    vec4<f32>(glyphPosition, 0.0, 1.0),
+    vec4<f32>(0.0),
+    isClipped
+  );
+  outputs.atlasUV = (glyphFrame.xy + corner * glyphSize) / vec2<f32>(textureDimensions(fontAtlasTexture));
+  outputs.color = textStorageStyleConfig.constantColor;
+  if (textStorageStyleConfig.useRowColors != 0u) {
+    outputs.color = unpackTextColor(textRowColors[rowStorageIndex]);
+  }
+  return outputs;
+}
+
+@fragment
+fn fragmentMain(inputs: FragmentInputs) -> @location(0) vec4<f32> {
+  let sampledAlpha = textureSample(fontAtlasTexture, fontAtlasTextureSampler, inputs.atlasUV).a;
+  var alpha = sampledAlpha;
+  if (textStorageStyleConfig.sdfThreshold >= 0.0) {
+    if (textStorageStyleConfig.sdfSmoothing > 0.0) {
+      alpha = smoothstep(
+        textStorageStyleConfig.sdfThreshold - textStorageStyleConfig.sdfSmoothing,
+        textStorageStyleConfig.sdfThreshold + textStorageStyleConfig.sdfSmoothing,
+        sampledAlpha
+      );
+    } else {
+      alpha = select(0.0, 1.0, sampledAlpha >= textStorageStyleConfig.sdfThreshold);
+    }
+  }
+  return vec4<f32>(inputs.color.rgb, inputs.color.a * alpha);
+}
+`;
